Extract common methods to util.py
